mpcbenchrunner/requirements.txt

# commonsnakes
pyyaml
attrs
cattrs
docker
../commonsnakes
python-dateutil
